step4: movie 개선 다형성, 변경 보호
