+++ src/chapter5/NoneDiscountMovie.ts
@@ -1,0 +1,12 @@
+import { Money } from "~src/chapter5/Money";
+import { Movie } from "~src/chapter5/Movie";
+
+export class NoneDiscountMovie extends Movie {
+  constructor(title: string, runningTime: number, fee: Money) {
+    super(title, runningTime, fee);
+  }
+
+  protected calculateDiscountAmount(): Money {
+    return Money.ZERO;
+  }
+}
